Did gradle refactoring
