+++ app/build.gradle
@@ -1,63 +1,17 @@
-/*
- * This file was generated by the Gradle 'init' task.
- *
- * This generated file contains a sample Scala application project to get you started.
- * For more details take a look at the 'Building Java & JVM projects' chapter in the Gradle
- * User Manual available at https://docs.gradle.org/7.2/userguide/building_java_projects.html
- */
-
 plugins {
-    // Apply the scala Plugin to add support for Scala.
     id 'scala'
-
-    // Apply the application plugin to add support for building a CLI application in Java.
     id 'application'
 }
 
 repositories {
-    // Use Maven Central for resolving dependencies.
     mavenCentral()
 }
 
 dependencies {
-    // Use Scala 2.13 in our library project
     implementation 'org.scala-lang:scala-library:2.13.6'
-
-    // This dependency is used by the application.
-    implementation 'com.google.guava:guava:30.1.1-jre'
-
     implementation 'com.lihaoyi:fastparse_2.13:2.3.3'
-    implementation 'org.reflections:reflections:0.10.2'
-    implementation 'org.scala-lang:scala-reflect:2.13.8'
-
-    // Use Scalatest for testing our library
-    testImplementation 'junit:junit:4.13.2'
-    testImplementation 'org.scalatest:scalatest_2.13:3.2.9'
-    testImplementation 'org.scalatestplus:junit-4-13_2.13:3.2.2.0'
-
-    // Need scala-xml at test runtime
-    testRuntimeOnly 'org.scala-lang.modules:scala-xml_2.13:1.2.0'
 }
 
 application {
-    // Define the main class for the application.
-    mainClass = 'scala.Main'
-}
-
-task runTests(type: JavaExec) {
-    classpath sourceSets.main.runtimeClasspath
-    main = "veritas.Veritas"
-}
-
-task runCoverage(type: JavaExec) {
-    classpath sourceSets.main.runtimeClasspath
-    main = "veritas.Veritas"
-
-    // If `extraArgs` exists, prepend it with "coverage" and set args to the result
-    // else just set args to "coverage"
-    if (project.hasProperty("extraArgs")) {
-        args(["coverage"] + (extraArgs.split(',') as List).flatten())
-    } else {
-        args("coverage")
-    }
+    mainClass = 'posharp.Main'
 }
