+++ app/src/main/scala/posharp/Util.scala
@@ -1,4 +1,4 @@
-package scala
+package posharp
 
 object Util {
   /**
