+++ docs/logo.png
