+++ gradle.properties
@@ -1,0 +1,5 @@
+# https://docs.gradle.org/current/userguide/command_line_interface.html#sec:command_line_performance
+org.gradle.configureondemand=true
+org.gradle.parallel=true
+org.gradle.caching=true
+org.gradle.daemon=true
