+++ veritas/settings.gradle
@@ -1,0 +1,4 @@
+rootProject.name = 'veritas'
+
+include ":app"
+project(":app").projectDir = file("../app")
