+++ veritas/src/main/scala/core/Constants.scala
@@ -1,0 +1,16 @@
+package core
+
+/**
+ * Holds a few constants that can potentially be referenced in multiple places in the code.
+ */
+object Constants {
+  /**
+   * Package name of the PoSharp implementation.
+   */
+  val POSHARP_PACKAGE = "posharp"
+
+  /**
+   * Path from the root of the project to the compiler's source code.
+   */
+  val POSHARP_PATH = s"app/src/main/scala/$POSHARP_PACKAGE/"
+}
