+++ veritas/src/main/scala/core/Veritas.scala
@@ -1,13 +1,14 @@
-package veritas
+package core
 
 import org.reflections.Reflections
 import org.reflections.scanners.Scanners.TypesAnnotated
 import org.reflections.util.ConfigurationBuilder
+import posharp.Main.writeToFile
+import posharp.{Parser, ToAssembly}
 
 import java.io.File
 import java.lang.reflect.Method
 import java.util.concurrent.{Executors, TimeUnit}
-import scala.Main.writeToFile
 import scala.io.AnsiColor._
 import scala.reflect.internal.util.ScalaClassLoader
 import scala.sys.process.Process
@@ -25,8 +26,8 @@ object Veritas {
    *
    * Command line arguments:
    * <ul>
-   *   <li>[0] - `coverage`: coverage is calculated and printed</li>
-   *   <li>[1] - `export`: coverage is exported in CodeCov JSON format</li>
+   * <li>[0] - `coverage`: coverage is calculated and printed</li>
+   * <li>[1] - `export`: coverage is exported in CodeCov JSON format</li>
    * </ul>
    *
    * Order matters!
@@ -153,14 +154,20 @@ object Veritas {
     if (calculateCoverage)
       cov.CalculateCoverage(exportCoverage)
 
-    // Delete all files created by writeToFile and the tests
+    deleteTestArtifacts()
+
+    exitCode
+  }
+
+  /**
+   * Deletes all files created by writeToFile and the tests.
+   */
+  private def deleteTestArtifacts(): Unit = {
     new File("compiled")
       .listFiles
       .filter(_.isFile)
       .filter(_.getName.contains("test"))
       .foreach(el => el.delete())
-
-    exitCode
   }
 
   /**
@@ -195,7 +202,7 @@ object Veritas {
 
     val tmp = Process(if (IsWindows()) {
       "wsl "
-    } + s"make TARGET_FILE=$fileName" else {
+    } + s"make -f ../Makefile TARGET_FILE=$fileName" else {
       ""
     }).!!
